Reduced Body Margins in the CSS.
Reduced the margin sizes in the CSS. While they 
look better on a large window, it breaks the 
experience in smaller windows.

For example, compare http://s.do.ugal.me/HRAs and 
http://s.do.ugal.me/HQfA - this is typically how
I work on RST files.

We could do something with media queries to make 
this work better in larger browsers, but I've not
looked at that option yet.